feat: add support for chat sessions (#167)
# What does this PR do?

today, to chat with a model, a user has to run one command per
completion. add --session which enables a user to have an interactive
chat session with the inference model. --session can be passed with or
without --message. If no --message is passed, the user is prompted to
give the first message.

This is useful as it also enables context to be saved between each
completion unlike today.

```
llama-stack-client inference chat-completion --session
>>> hi whats up!
Assistant> Not much! How's your day going so far? Is there something I can help you with or would you like to chat?
>>> what color is the sky?
Assistant> The color of the sky can vary depending on the time of day and atmospheric conditions. Here are some common colors you might see:

* During the daytime, when the sun is overhead, the sky typically appears blue.
* At sunrise and sunset, the sky can take on hues of red, orange, pink, and purple due to the scattering of light by atmospheric particles.
* On a clear day with no clouds, the sky can appear a bright blue, often referred to as "cerulean."
* In areas with high levels of pollution or dust, the sky can appear more hazy or grayish.
* At night, the sky can be dark and black, although some stars and moonlight can make it visible.

So, what's your favorite color of the sky?
>>>
```


## Test Plan

tested locally with and without --message

Signed-off-by: Charlie Doern <cdoern@redhat.com>