fix: Don't retry for non-recoverable server http errors
This is specifically addressing the issue where server returning Not
Implemented (code 501) would receive two more attempts for the same
request, even though there's no reason to expect it to serve the request
any better on further attempts.

This patch reduces the number of >=500 codes that would be restarted to
those where there seems to be a chance of recover on further attempts.
These codes are now explicitly listed instead of broad >=500 filter.

For all possible server codes, please consult e.g. here:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status#server_error_responses

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>